requirements.txt

streamlit
pypdf
python-docx
sentence-transformers
transformers
faiss-cpu
langchain
torch
